Serialize event handling per pod
Before this change all pod events are handling concurrently. This means
that multiple events for the same pod can be handled at the same time
and end up in a race where they may be processed in the incorrect order.

An easy way to reproduce this issue is to fire up VK, add some pods, then suspend it
(ctrl+z) for a minute or two, then resume. A blast of events will come
from k8s and cause out of order handling. This will cause duplicate pods
to be created in the provider that can never be cleaned up (by VK).

This change uses a mutex keyed on pod namespace+name to ensure that only
one event is handled at a time and as an added check to ensure that the
pod does not already exist in the resource manager before triggering a
pod create.