Introduce Soft Deletes to allow for Terminal Status Updates
This introduces the concept of soft deletes, so that the
virtual-kubelet can make a call to API server to update
the pod and set the terminal state.

The mechanism of action is that if soft deletes are enabled
(by the provider), the kubelet will then not actually delete
the pod from apiserver, but instead set an annotation
indicating that it has been deleted. Then, on subsequent
updates, it will not make the callback to the provider
to repeat deleting the pod.

This also makes some changes with how updates are processed
with the asynchronous notifier. Instead of the system going
like this:

 Provider               Server
    |                     |
    |   Notify Update     |
    |-------------------->|
    |                     |
    |<--------------------|
    |    GetPodStatus     |
    |                     |
    |-------------------->|
    |  Pod + Pod Status   |
    |                     |

The pod status updates are shipped with the notifications.

The mock provider has been updated to use this interface.