Add "TerminatePod" to lifecycle interface
This splits the termination of a pod from the deletion of it.
It allows providers to update the final pod status before deleting it is
deleted form k8s.

Pod is terminated if pod.Phase == Running and pod.DeletionTimestamp is
not nil.

When the provider is ready for the pod to be deleted, the pod status
should be udpated so it's not "Running", but whatever is appropriate for
the exist statuses of the containers.
Once the status is updated a new event will be triggered from k8s which
will trigger us to call `DeletePod` as before.