Have NotifyPods store the pod status in a map
We introduce a map that can be used to store the pod status. In this,
we do not need to call GetPodStatus immediately after NotifyPods
is called. Instead, we stash the pod passed via notifypods
as in a map we can access later. In addition to this, for legacy
providers, the logic to merge the pod, and the pod status is
hoisted up to the loop.

It prevents leaks by deleting the entry in the map as soon
as the pod is deleted from k8s.

In adddition, it introduces a map that's used to store pods at
termination. This prevents duplicate pod terminations, as well
as allows us to avoid the infinite loop of pod termination
status updates, as a naive provider would make a pod status
update everytime DeletePod was called. This in turn would
trigger an update in apiserver, resulting in an event,
and subsequently an infinite loop.