Add GetStatsSummary to PodHandlerConfig
If both the metrics routes and the pod routes are attached to the same
mux with the pattern "/", it will panic. Instead, add the stats handler
function to PodHandlerConfig and set up the route if it is not nil.

(cherry picked from commit 56b248c8549833c10e0d7b966c4b8086555879d3)
Signed-off-by: Brian Goff <cpuguy83@gmail.com>