Queue: Use sync.Cond instead of wake channel
The reason the channel was used is because it enables us to select on
multiple things at once to wait for.
Really we don't need to wait for multiple things to happen since we can
use a condition variable to wakeup for any number of reasons.
Since the queue is already waiting for the `ctx.Done()` we can use that
to signal to workers to wake up and exit.